requirements.txt
videodb
streamlit
google-generativeai
python-dotenv